Update to GCC 15.0.1

+++ build.sh
@@ -6,6 +6,9 @@ declare -r workdir="${PWD}"
 
 declare -r revision="$(git rev-parse --short HEAD)"
 
+declare -r toolchain_directory='/tmp/atar'
+declare -r share_directory="${toolchain_directory}/usr/local/share/atar"
+
 declare -r gmp_tarball='/tmp/gmp.tar.xz'
 declare -r gmp_directory='/tmp/gmp-6.3.0'
 
@@ -16,11 +19,9 @@ declare -r mpc_tarball='/tmp/mpc.tar.gz'
 declare -r mpc_directory='/tmp/mpc-1.3.1'
 
 declare -r binutils_tarball='/tmp/binutils.tar.xz'
-declare -r binutils_directory='/tmp/binutils-2.43'
+declare -r binutils_directory='/tmp/binutils-with-gold-2.44'
 
 declare -r lld_tarball='/tmp/lld.tar.xz'
-
-declare -r toolchain_directory='/tmp/atar'
 
 declare gcc_directory=''
 
@@ -33,10 +34,10 @@ function setup_gcc_source() {
 	
 	declare -r tgt
 	
-	if [ "${tgt}" = 'hppa' ] || [ "${tgt}" = 'alpha' ] || [ "${tgt}" = 'amd64' ] || [ "${tgt}" = 'i386' ]; then
-		gcc_version='14'
-		gcc_directory='/tmp/gcc-14.2.0'
-		gcc_url='https://ftp.gnu.org/gnu/gcc/gcc-14.2.0/gcc-14.2.0.tar.xz'
+	if [ "${tgt}" = 'hppa-unknown-openbsd' ] || [ "${tgt}" = 'alpha-unknown-openbsd' ] || [ "${tgt}" = 'x86_64-unknown-openbsd' ] || [ "${tgt}" = 'i386-unknown-openbsd' ]; then
+		gcc_version='15'
+		gcc_directory='/tmp/gcc-master'
+		gcc_url='https://github.com/gcc-mirror/gcc/archive/refs/heads/master.tar.gz'
 	else
 		gcc_version='11'
 		gcc_directory='/tmp/gcc-11.2.0'
@@ -62,11 +63,18 @@ function setup_gcc_source() {
 				patch --directory="${gcc_directory}" --strip='0' --input="${name}"
 			done
 			
+			sed --in-place 's/HAVE_IFUNC/(HAVE_IFUNC \&\& !defined(__arm__))/g' "${gcc_directory}/libatomic/libatomic_i.h"
+			
 			patch --directory="${gcc_directory}" --strip='1' --input="${workdir}/patches/0001-Thats-not-openbsd.patch"
 			patch --directory="${gcc_directory}" --strip='1' --input="${workdir}/patches/0001-Add-libversions.patch"
-		elif (( gcc_version >= 14 )); then
+		elif (( gcc_version >= 15 )); then
 			patch --directory="${gcc_directory}" --strip='1' --input="${workdir}/patches/0001-Fix-libatomic-build-with-newer-GCC-versions.patch"
 			patch --directory="${gcc_directory}" --strip='1' --input="${workdir}/patches/0001-Disable-libfunc-support-for-hppa-unknown-openbsd.patch"
+			patch --directory="${gcc_directory}" --strip='1' --input="${workdir}/submodules/obggcc/patches/0001-Fix-libgcc-build-on-arm.patch"
+			patch --directory="${gcc_directory}" --strip='1' --input="${workdir}/submodules/obggcc/patches/0001-Change-the-default-language-version-for-C-compilatio.patch"
+			patch --directory="${gcc_directory}" --strip='1' --input="${workdir}/submodules/obggcc/patches/0001-Turn-Wimplicit-int-back-into-an-warning.patch"
+			patch --directory="${gcc_directory}" --strip='1' --input="${workdir}/submodules/obggcc/patches/0001-Turn-Wint-conversion-back-into-an-warning.patch"
+			patch --directory="${gcc_directory}" --strip='1' --input="${workdir}/submodules/obggcc/patches/0001-Revert-GCC-change-about-turning-Wimplicit-function-d.patch"
 		fi
 		
 		touch "${gcc_directory}/patched"
@@ -74,7 +82,7 @@ function setup_gcc_source() {
 	
 }
 
-declare optflags='-Os'
+declare optflags=''
 declare -r linkflags='-Wl,-s'
 
 declare -r max_jobs="$(($(nproc) * 17))"
@@ -91,9 +99,7 @@ if [ "${build_type}" == 'native' ]; then
 	is_native='1'
 fi
 
-declare OBGGCC_TOOLCHAIN='/tmp/obggcc-toolchain'
 declare CROSS_COMPILE_TRIPLET=''
-
 declare cross_compile_flags=''
 
 if ! (( is_native )); then
@@ -117,11 +123,11 @@ if ! [ -f "${mpc_tarball}" ]; then
 fi
 
 if ! [ -f "${binutils_tarball}" ]; then
-	wget --no-verbose 'https://ftp.gnu.org/gnu/binutils/binutils-2.43.tar.xz' --output-document="${binutils_tarball}"
+	wget --no-verbose 'https://ftp.gnu.org/gnu/binutils/binutils-with-gold-2.44.tar.xz' --output-document="${binutils_tarball}"
 	tar --directory="$(dirname "${binutils_directory}")" --extract --file="${binutils_tarball}"
 	
-	patch --directory="${binutils_directory}" --strip='1' --input="${workdir}/patches/0001-Revert-gold-Use-char16_t-char32_t-instead-of-uint16_.patch"
-	patch --directory="${binutils_directory}" --strip='1' --input="${workdir}/patches/0001-Disable-warning-regarding-exec-stack.patch"
+	patch --directory="${binutils_directory}" --strip='1' --input="${workdir}/submodules/obggcc/patches/0001-Revert-gold-Use-char16_t-char32_t-instead-of-uint16_.patch"
+	patch --directory="${binutils_directory}" --strip='1' --input="${workdir}/submodules/obggcc/patches/0001-Disable-annoying-linker-warnings.patch"
 fi
 
 if ! [ -f "${lld_tarball}" ]; then
@@ -140,7 +146,6 @@ fi
 [ -d "${gmp_directory}/build" ] || mkdir "${gmp_directory}/build"
 
 cd "${gmp_directory}/build"
-rm --force --recursive ./*
 
 ../configure \
 	--prefix="${toolchain_directory}" \
@@ -157,7 +162,6 @@ make install
 [ -d "${mpfr_directory}/build" ] || mkdir "${mpfr_directory}/build"
 
 cd "${mpfr_directory}/build"
-rm --force --recursive ./*
 
 ../configure \
 	--prefix="${toolchain_directory}" \
@@ -175,7 +179,6 @@ make install
 [ -d "${mpc_directory}/build" ] || mkdir "${mpc_directory}/build"
 
 cd "${mpc_directory}/build"
-rm --force --recursive ./*
 
 ../configure \
 	--prefix="${toolchain_directory}" \
@@ -193,58 +196,28 @@ make install
 [ -d "${binutils_directory}/build" ] || mkdir "${binutils_directory}/build"
 
 declare -r targets=(
-	'armv7'
-	'amd64'
-	'riscv64'
-	'arm64'
-	'powerpc64'
-	'macppc'
-	'sparc64'
-	'octeon'
-	'loongson'
-	'hppa'
-	'alpha'
-	'i386'
+	'x86_64-unknown-openbsd'
+	'i386-unknown-openbsd'
+	'alpha-unknown-openbsd'
+	'arm-unknown-openbsd'
+	'hppa-unknown-openbsd'
+	'aarch64-unknown-openbsd'
+	'powerpc-unknown-openbsd'
+	'powerpc64-unknown-openbsd'
+	'sparc64-unknown-openbsd'
+	'mips64-unknown-openbsd'
+	'mips64el-unknown-openbsd'
+	'riscv64-unknown-openbsd'
 )
 
-for target in "${targets[@]}"; do
-	case "${target}" in
-		armv7)
-			declare triplet='arm-unknown-openbsd';;
-		arm64)
-			declare triplet='aarch64-unknown-openbsd';;
-		macppc)
-			declare triplet='powerpc-unknown-openbsd';;
-		powerpc64)
-			declare triplet='powerpc64-unknown-openbsd';;
-		sparc64)
-			declare triplet='sparc64-unknown-openbsd';;
-		octeon)
-			declare triplet='mips64-unknown-openbsd';;
-		loongson)
-			declare triplet='mips64el-unknown-openbsd';;
-		riscv64)
-			declare triplet='riscv64-unknown-openbsd';;
-		amd64)
-			declare triplet='x86_64-unknown-openbsd';;
-		i386)
-			declare triplet='i386-unknown-openbsd';;
-		hppa)
-			declare triplet='hppa-unknown-openbsd';;
-		alpha)
-			declare triplet='alpha-unknown-openbsd';;
-	esac
-	
-	wget --no-verbose "https://mirrors.ucr.ac.cr/pub/OpenBSD/7.0/${target}/base70.tgz" --output-document='/tmp/base.tgz'
-	wget --no-verbose "https://mirrors.ucr.ac.cr/pub/OpenBSD/7.0/${target}/comp70.tgz" --output-document='/tmp/comp.tgz'
-	
+for triplet in "${targets[@]}"; do
 	cd "${binutils_directory}/build"
 	rm --force --recursive ./*
 	
 	declare extra_binutils_flags=''
 	declare require_lld='0'
 	
-	if [ "${target}" = 'armv7' ] || [ "${target}" = 'arm64' ]; then
+	if [ "${triplet}" = 'arm-unknown-openbsd' ] || [ "${triplet}" = 'aarch64-unknown-openbsd' ]; then
 		require_lld='1'
 	fi
 	
@@ -269,6 +242,25 @@ for target in "${targets[@]}"; do
 	make all --jobs > /dev/null
 	make install
 	
+	cd "$(mktemp --directory)"
+	
+	declare sysroot_url="https://github.com/AmanoTeam/openbsd-sysroot/releases/latest/download/${triplet}.tar.xz"
+	declare sysroot_file="${PWD}/${triplet}.tar.xz"
+	declare sysroot_directory="${PWD}/${triplet}"
+	
+	wget \
+		--no-verbose \
+		--output-document="${sysroot_file}" \
+		"${sysroot_url}"
+	
+	tar \
+		--extract \
+		--file="${sysroot_file}"
+	
+	cp --recursive "${sysroot_directory}" "${toolchain_directory}"
+	
+	rm --force --recursive ./*
+	
 	cd "${toolchain_directory}/bin"
 	
 	ln --symbolic './ld.lld' "./${triplet}-ld.lld"
@@ -277,52 +269,43 @@ for target in "${targets[@]}"; do
 		ln --symbolic './ld.lld' "./${triplet}-ld"
 	fi
 	
-	tar --directory="${toolchain_directory}/${triplet}" --strip=2 --extract --file='/tmp/base.tgz' './usr/lib' './usr/include'
-	tar --directory="${toolchain_directory}/${triplet}" --strip=2 --extract --file='/tmp/comp.tgz' './usr/lib' './usr/include'
-	
-	cd "${toolchain_directory}/${triplet}/lib"
-	
-	while read source; do
-		IFS='.' read -ra parts <<< "${source}"
-		
-		declare name="${parts[1]}"
-		declare destination="${name#/}.so"
-		
-		ln --symbolic "${source}" "./${destination}"
-	done <<< "$(find '.' -type 'f' -name 'lib*.so.*')"
-	
-	setup_gcc_source "${target}"
+	setup_gcc_source "${triplet}"
 	
 	cd "${gcc_directory}/build"
 	
 	rm --force --recursive ./*
 	
 	declare extra_configure_flags=''
-	declare supports_lto='0'
+	declare have_lto='0'
+	declare have_disable_fixincludes='0'
 	
-	if [ "${target}" == 'hppa' ]; then
+	if [ "${triplet}" = 'hppa-unknown-openbsd' ]; then
 		extra_configure_flags+='--disable-libstdcxx '
 	fi
 	
-	if [ "${target}" == 'hppa' ] || [ "${target}" == 'alpha' ] || [ "${target}" == 'amd64' ] || [ "${target}" == 'i386' ]; then
-		supports_lto='1'
+	if [ "${triplet}" = 'hppa-unknown-openbsd' ] || [ "${triplet}" = 'alpha-unknown-openbsd' ] || [ "${triplet}" = 'x86_64-unknown-openbsd' ] || [ "${triplet}" = 'i386-unknown-openbsd' ]; then
+		have_disable_fixincludes='1'
 	fi
 	
-	if (( supports_lto )); then
+	if [ "${triplet}" = 'hppa-unknown-openbsd' ] || [ "${triplet}" = 'alpha-unknown-openbsd' ] || [ "${triplet}" = 'x86_64-unknown-openbsd' ] || [ "${triplet}" = 'i386-unknown-openbsd' ]; then
+		have_lto='1'
+	fi
+	
+	if (( have_disable_fixincludes )); then
+		extra_configure_flags+='--disable-fixincludes '
+	fi
+	
+	if (( have_lto )); then
 		extra_configure_flags+='--enable-lto '
 	else
 		extra_configure_flags+='--disable-lto '
 	fi
 	
-	if [ "${target}" = 'armv7' ]; then
-		extra_configure_flags+='--disable-libatomic '
-	fi
-	
 	# The compiler for powerpc64 breaks if compiled with -Os
-	if [ "${target}" == 'powerpc64' ]; then
-		optflags='-O2'
+	if [ "${triplet}" = 'powerpc64-unknown-openbsd' ]; then
+		optflags='-w -O2'
 	else
-		optflags='-Os'
+		optflags='-w -Os'
 	fi
 	
 	../configure \
@@ -334,7 +317,7 @@ for target in "${targets[@]}"; do
 		--with-mpfr="${toolchain_directory}" \
 		--with-bugurl='https://github.com/AmanoTeam/Atar/issues' \
 		--with-gcc-major-version-only \
-		--with-pkgversion="Atar v0.5-${revision}" \
+		--with-pkgversion="Atar v0.6-${revision}" \
 		--with-sysroot="${toolchain_directory}/${triplet}" \
 		--with-native-system-header-dir='/include' \
 		--enable-__cxa_atexit \
@@ -343,11 +326,16 @@ for target in "${targets[@]}"; do
 		--enable-gnu-indirect-function \
 		--enable-gnu-unique-object \
 		--enable-libstdcxx-backtrace \
-		--enable-plugin \
 		--enable-shared \
 		--enable-threads='posix' \
 		--enable-languages='c,c++' \
+		--enable-cpp \
+		--enable-default-pie \
+		--enable-standard-branch-protection \
+		--enable-wchar_t \
 		--without-headers \
+		--disable-plugin \
+		--disable-libsanitizer \
 		--disable-bootstrap \
 		--disable-libgomp \
 		--disable-libmudflap \
@@ -357,20 +345,21 @@ for target in "${targets[@]}"; do
 		--disable-nls \
 		--disable-tls \
 		--disable-werror \
-		--enable-cpp \
-		--enable-default-pie \
-		--enable-standard-branch-protection \
-		--enable-wchar_t \
 		${cross_compile_flags} \
 		${extra_configure_flags} \
 		am_cv_func_iconv=no \
 		ac_cv_header_magic_h=no \
 		CFLAGS="${optflags}" \
 		CXXFLAGS="${optflags}" \
-		LDFLAGS="-Wl,-rpath-link,${OBGGCC_TOOLCHAIN}/${CROSS_COMPILE_TRIPLET}/lib ${linkflags}"
+		LDFLAGS="${linkflags}"
 	
 	declare CFLAGS_FOR_TARGET="${optflags} ${linkflags}"
-	declare CXXFLAGS_FOR_TARGET="${optflags} ${linkflags} -nostdinc++ -fpermissive"
+	declare CXXFLAGS_FOR_TARGET="${optflags} ${linkflags} -fpermissive"
+	
+	# https://gcc.gnu.org/bugzilla/show_bug.cgi?id=80196#c12
+	if ! (( is_native )); then
+		CXXFLAGS_FOR_TARGET+=' -nostdinc++'
+	fi
 	
 	LD_LIBRARY_PATH="${toolchain_directory}/lib" PATH="${PATH}:${toolchain_directory}/bin" make \
 		CFLAGS_FOR_TARGET="${CFLAGS_FOR_TARGET}" \
@@ -380,24 +369,24 @@ for target in "${targets[@]}"; do
 	
 	cd "${toolchain_directory}/${triplet}/bin"
 	
-	for name in *; do
-		rm "${name}"
-		ln --symbolic "../../bin/${triplet}-${name}" "${name}"
-	done
-	
 	ln --symbolic '../../bin/ld.lld' 'ld.lld'
 	
 	if (( require_lld )); then
 		ln --symbolic '../../bin/ld.lld' 'ld'
 	fi
 	
-	rm --recursive "${toolchain_directory}/share"
-	rm --recursive "${toolchain_directory}/lib/gcc/${triplet}/"*"/include-fixed"
+	if ! (( have_disable_fixincludes )); then
+		rm --recursive "${toolchain_directory}/lib/gcc/${triplet}/"*"/include-fixed"
+	fi
 	
 	patchelf --add-rpath '$ORIGIN/../../../../lib' "${toolchain_directory}/libexec/gcc/${triplet}/"*"/cc1"
 	patchelf --add-rpath '$ORIGIN/../../../../lib' "${toolchain_directory}/libexec/gcc/${triplet}/"*"/cc1plus"
 	
-	if (( supports_lto )); then
+	if (( have_lto )); then
 		patchelf --add-rpath '$ORIGIN/../../../../lib' "${toolchain_directory}/libexec/gcc/${triplet}/"*"/lto1"
 	fi
 done
+
+mkdir --parent "${share_directory}"
+
+cp --recursive "${workdir}/tools/dev/"* "${share_directory}"
